kreirana metoda dajUkupnuCijenuArtikala() u klasi Korpa

+++ src/ba/unsa/etf/rpr/Artikl.java
@@ -2,7 +2,7 @@ package ba.unsa.etf.rpr;
 
 public class Artikl {
     private String naziv;
-    private double cijena;
+    private int cijena;
     private String kod;
 
     public Artikl(String naziv, int cijena, String kod) {
@@ -22,7 +22,6 @@ public class Artikl {
     public String getKod() {
         return kod;
     }
-
 
 }
